+++ lib/kitchen/suite.rb
@@ -27,55 +27,75 @@ module Kitchen
     # @return [String] logical name of this suite
     attr_reader :name
 
-    # @return [Array] Array of Chef run_list items
-    attr_reader :run_list
-
-    # @return [Hash] Hash of Chef node attributes
-    attr_reader :attributes
-
     # @return [Array] Array of names of excluded platforms
     attr_reader :excludes
-
-    # @return [String] local path to the suite's data bags, or nil if one does
-    #   not exist
-    attr_reader :data_bags_path
-
-    # @return [String] local path to the suite's encrypted data bag secret
-    #   key path, or nil if one does not exist
-    attr_reader :encrypted_data_bag_secret_key_path
-
-    # @return [String] local path to the suite's roles, or nil if one does
-    #   not exist
-    attr_reader :roles_path
 
     # Constructs a new suite.
     #
     # @param [Hash] options configuration for a new suite
     # @option options [String] :name logical name of this suit (**Required**)
-    # @option options [String] :run_list Array of Chef run_list items
-    #   (**Required**)
-    # @option options [Hash] :attributes Hash of Chef node attributes
     # @option options [String] :excludes Array of names of excluded platforms
-    # @option options [String] :data_bags_path path to data bags
-    # @option options [String] :roles_path path to roles
-    # @option options [String] :encrypted_data_bag_secret_key_path path to
-    #   secret key file
     def initialize(options = {})
+      options = options.dup
       validate_options(options)
 
-      @name = options[:name]
-      @run_list = options[:run_list]
-      @attributes = options[:attributes] || Hash.new
-      @excludes = options[:excludes]     || Array.new
-      @data_bags_path = options[:data_bags_path]
-      @roles_path = options[:roles_path]
-      @encrypted_data_bag_secret_key_path = options[:encrypted_data_bag_secret_key_path]
+      @name = options.delete(:name)
+      @excludes = Array(options[:excludes])
+      @data = options
+    end
+
+    # Extra suite methods used for accessing Chef data such as a run list,
+    # node attributes, etc.
+    module Cheflike
+
+      # @return [Array] Array of Chef run_list items
+      def run_list
+        Array(data[:run_list])
+      end
+
+      # @return [Hash] Hash of Chef node attributes
+      def attributes
+        data[:attributes] || Hash.new
+      end
+
+      # @return [String] local path to the suite's data bags, or nil if one
+      #   does not exist
+      def data_bags_path
+        data[:data_bags_path]
+      end
+
+      # @return [String] local path to the suite's encrypted data bag secret
+      #   key path, or nil if one does not exist
+      def encrypted_data_bag_secret_key_path
+        data[:encrypted_data_bag_secret_key_path]
+      end
+
+      # @return [String] local path to the suite's roles, or nil if one does
+      #   not exist
+      def roles_path
+        data[:roles_path]
+      end
+
+      # @return [String] local path to the suite's nodes, or nil if one does
+      #   not exist
+      def nodes_path
+        data[:nodes_path]
+      end
+    end
+
+    # Extra suite methods used for accessing Puppet data such as a manifest.
+    module Puppetlike
+
+      def manifest
+      end
     end
 
     private
 
+    attr_reader :data
+
     def validate_options(opts)
-      [:name, :run_list].each do |k|
+      [:name].each do |k|
         raise ClientError, "Suite#new requires option :#{k}" if opts[k].nil?
       end
     end